Adds QASM for alternate "b" compilations, nontrivial 2Q gate counts, and DRB latex comment.
Adds .qasm files in circuits/ionq/ that give QASM circuits for
alternate compilations of AQ circuits that were used in some of the
executions.  The results/data_AQcircuits.json is updated to include
a compilation_suffix column to indicate when these "b" compilations
were used.

This commit also adds a column to the AQ data (data_AQcircuits.json)
indicating the number of nontrivial 2-qubit gates in each circuit,
even though in executions leading to the data these trivial gates were
performed.

Finally, this commit adds a comment to a cell in the DRB figure
generation notebook indicating use of latex in matplotlib is optional.